Move POPhoneNumberFieldState
