+++ ui-core/src/main/kotlin/com/processout/sdk/ui/core/state/POPhoneNumberFieldState.kt
@@ -1,13 +1,10 @@
-package com.processout.sdk.ui.core.component.field.phone
+package com.processout.sdk.ui.core.state
 
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.runtime.Immutable
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.input.VisualTransformation
 import com.processout.sdk.ui.core.annotation.ProcessOutInternalApi
-import com.processout.sdk.ui.core.state.POAvailableValue
-import com.processout.sdk.ui.core.state.POImmutableList
-import com.processout.sdk.ui.core.state.POInputFilter
 
 /** @suppress */
 @ProcessOutInternalApi
