+++ ui/src/main/kotlin/com/processout/sdk/ui/napm/v2/NativeAlternativePaymentViewModelState.kt
@@ -1,9 +1,9 @@
 package com.processout.sdk.ui.napm.v2
 
 import androidx.compose.runtime.Immutable
-import com.processout.sdk.ui.core.component.field.phone.POPhoneNumberFieldState
 import com.processout.sdk.ui.core.state.POActionState
 import com.processout.sdk.ui.core.state.POImmutableList
+import com.processout.sdk.ui.core.state.POPhoneNumberFieldState
 import com.processout.sdk.ui.shared.state.ConfirmationDialogState
 import com.processout.sdk.ui.shared.state.FieldState
